refactor(mobile): replace ValueNotifier with cubit subscription in reasoning modal
